Move camera modules to parallax/camera/ folder
