+++ parallax/cameras/camera.py
@@ -8,7 +8,7 @@ import threading
 import time
 import cv2
 import numpy as np
-from parallax.screens.camera_base_binding import BaseCamera
+from parallax.cameras.camera_base_binding import BaseCamera
 
 # Initialize the logger
 logger = logging.getLogger(__name__)
