+++ parallax/model.py
@@ -8,7 +8,7 @@ their initialization, configuration, and transformations between local and globa
 """
 from collections import OrderedDict
 from PyQt5.QtCore import QObject
-from parallax.screens.camera import MockCamera, PySpinCamera, close_cameras, list_cameras
+from parallax.cameras.camera import MockCamera, PySpinCamera, close_cameras, list_cameras
 from parallax.stages.stage_listener import Stage, StageInfo
