+++ parallax/reticle_detection/reticle_detect_manager.py
@@ -11,7 +11,7 @@ import cv2
 import numpy as np
 from PyQt5.QtCore import QObject, QThread, pyqtSignal
 
-from parallax.screens.calibration_camera import CalibrationCamera
+from parallax.cameras.calibration_camera import CalibrationCamera
 from parallax.reticle_detection.mask_generator import MaskGenerator
 from parallax.reticle_detection.reticle_detection import ReticleDetection
 from parallax.reticle_detection.reticle_detection_coords_interests import ReticleDetectCoordsInterest
