+++ parallax/screens/axis_filter.py
@@ -10,7 +10,7 @@ import time
 import cv2
 import numpy as np
 from PyQt5.QtCore import QObject, QThread, pyqtSignal
-from .calibration_camera import CalibrationCamera
+from parallax.cameras.calibration_camera import CalibrationCamera
 
 # Set logger name
 logger = logging.getLogger(__name__)
